Move utility from ToggleGroupControl to utils.
